forward: refactor to exclusively use channel-based implementation
